requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
Pillow>=8.4.0
scikit-learn>=0.24.2
pyyaml>=5.4.1
tensorboard>=2.7.0
